FEAT: 동아리 관련 repository 업데이트 및 추가
- `ClubEventRepository` 삭제
- `ClubGalPhotoRepository` 추가
- `ClubMemberRepository`, `ClubPhotoRepository`, `ClubRepository`, `ClubUniversityRepository`, `ClubUserRepository`의 ID 타입을 Long에서 Integer로 변경
- `ClubPhotoRepository`에 `findByClubIdAndDeletedAtIsNull` 메서드 추가
- `UniversityRepository` 추가 (University 도메인)

+++ src/main/java/com/jandi/band_backend/club/repository/ClubMemberRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ClubMemberRepository extends JpaRepository<ClubMember, Long> {
-    List<ClubMember> findByClubId(Long clubId);
-    Optional<ClubMember> findByClubIdAndUserId(Long clubId, Long userId);
-    void deleteByClubIdAndUserId(Long clubId, Long userId);
+public interface ClubMemberRepository extends JpaRepository<ClubMember, Integer> {
+    List<ClubMember> findByClubId(Integer clubId);
+    Optional<ClubMember> findByClubIdAndUserId(Integer clubId, Integer userId);
+    void deleteByClubIdAndUserId(Integer clubId, Integer userId);
 }
